separate MainActivityUI class with MainActivity
